refactored component-level abstractions out
