+++ requirements.txt
@@ -1,6 +1,4 @@
 pytest>=7.0.0
-black>=23.0.0
-flake8>=6.0.0
 numpy
 scipy
 matplotlib
